pyproject.toml
[project]
name = "minimal-agent"
version = "2.0.0"
description = "Dual-framework AI agent platform (roo-agent + open-agent)"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "click>=8.0",
    "prompt-toolkit>=3.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.0",
    "aiosqlite>=0.19",
    "tiktoken>=0.5",
    "pydantic>=2.0",
]

[project.optional-dependencies]
pdf = ["pymupdf>=1.23"]
api = [
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "sse-starlette>=1.8",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
]

[project.scripts]
mini-agent = "minimal_agent.cli:main"
roo-agent = "roo_agent.cli.app:main"
open-agent = "open_agent.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/roo_agent", "src/open_agent", "src/minimal_agent"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
